added motif representation table in csv
This file contains a schematic representation of the Z-score measuring motif representation. The results are generated from an ensamble of 50 random graphs obtained by the swap algorithm that preserves the degree in and out of each node. For each of these 50 graphs, 100 * E swaps were performed where E is the number of arcs.
Each cell has one of these three symbols:
🔺means over-represented motif
- means the number of that motif is 0 in both the original and the randomized dataset
⬇means under-represented motif

Translated with DeepL.com (free version)